# docs/openapi/management.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Heidi Management API",
    "description" : "Administrative API for Cockpit and management automation. This is not the API used to run wallet processes.",
    "version" : "1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/management/v1/identities/tenants/{tenantId}/{id}/credential-encryption" : {
      "get" : {
        "tags" : [ "issuer-controller" ],
        "operationId" : "getCredentialEncryption",
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuerCredentialEncryption"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "issuer-controller" ],
        "operationId" : "updateCredentialEncryption",
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssuerCredentialEncryption"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuerCredentialEncryption"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CredentialEncryptionKey" : {
        "type" : "object",
        "properties" : {
          "keyId" : {
            "type" : "string"
          },
          "keyUri" : {
            "type" : "string"
          },
          "algorithm" : {
            "type" : "string"
          },
          "publicJwk" : {
            "type" : "string"
          },
          "providerEndpoint" : {
            "type" : "string"
          },
          "providerAuthenticationMode" : {
            "type" : "string"
          },
          "keyAlgorithm" : {
            "type" : "string"
          },
          "trustSystem" : {
            "type" : "string",
            "enum" : [ "Switzerland", "EUDI", "Custom", "OIDF", "Default" ]
          }
        }
      },
      "IssuerCredentialEncryption" : {
        "type" : "object",
        "properties" : {
          "requestAlgValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "requestEncValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "requestZipValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "responseAlgValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "responseEncValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "responseZipValues" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "requestEncryptionRequired" : {
            "type" : "boolean"
          },
          "responseEncryptionRequired" : {
            "type" : "boolean"
          },
          "requestKeys" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CredentialEncryptionKey"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "platformSession" : {
        "type" : "http",
        "description" : "Authenticated platform session.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}